refactor(Groceries) Renamed manipulation file
