+++ challenge_storing_groceries/scripts/test_manipulation
@@ -2,7 +2,7 @@
 
 import rospy
 from robot_smach_states.util.startup import startup
-from challenge_storing_groceries.manipulate_single_item import ManipulateMachine
+from challenge_storing_groceries.manipulate_machine import ManipulateMachine
 
 
 if __name__ == '__main__':
