+++ challenge_storing_groceries/src/challenge_storing_groceries/manipulate_machine.py
